backend/requirements.txt
# ClimaScope – Backend Dependencies (Multi-User AI Platform)
# Install with:
#   pip install -r requirements.txt
#
# Run with:
#   uvicorn app.main:app --host 0.0.0.0 --port 8000 --reload
#
# Environment variables needed:
#   MONGODB_URL=mongodb://localhost:27017
#   DATABASE_NAME=climascope
#   JWT_SECRET_KEY=your_secret_key_here
#   CLIMASCOPE_CORS_ORIGINS=http://localhost:5173

# ── Framework ─────────────────────────────────────────────────────────────────
fastapi==0.104.1
uvicorn[standard]==0.24.0

# ── Database ──────────────────────────────────────────────────────────────────
pymongo==4.6.0
motor==3.3.2

# ── Authentication ─────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0

# ── Validation ────────────────────────────────────────────────────────────────
pydantic==2.5.0
python-multipart==0.0.6

# ── AI / Machine Learning ─────────────────────────────────────────────────────
scikit-learn==1.4.0
google-generativeai==0.8.3
joblib==1.3.2
numpy==1.26.0
pandas==2.2.0
matplotlib==3.8.0
seaborn==0.13.0
openpyxl==3.1.2

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
httpx==0.27.2
jinja2==3.1.2
aiofiles==23.2.1
twilio==9.3.0
